pulling in public annos as for dashboard
This reveals that there are a now quite a few duplicates,
I'm guessing these were the ones that intermitently failed
due to the hypothesis bug and somehow they magically were
reinserted, though some were from the Nov 4 2016 adventure,
so those are a different issue. Will need to write a dedupe
script when we get the chance.