long flush times when remote solved
The issue here is that when doing bulk inserts even core sqlalchemy is
extremely slow when calling executemany for reasons that I do not
entirely understand. The solution is to construct the bulk equivalent
using makeParamsValues so that only a single sql statement is issued
per table that needs to be updated and handle the sequencing yourself.
psycopg2 was spending a huge amount of time on cursor.execute when
remote, probably due to the round trip lag time for each call.